moved cpp unit tests; working on data structure.

+++ src/common_cpp/Recon/SingleStation/SEDigit.hh
@@ -69,7 +69,7 @@ class SEDigit {
 
   double get_npe()     const { return _npe; }
 
-  void set_adc(double adc) { _adc = adc; }
+  void set_adc(int adc) { _adc = adc; }
 
   double get_adc()     const { return _adc; }
